+++ CHANGELOG.md
@@ -5,7 +5,7 @@ CHANGE LOG
 ## V2.0 Beta 1 (Upcoming - July 2014)
 
 * Upgrade to Laravel ~4.1
-* Moved to parsedown
+* Moved to parsedown extra
 * Refactoring
 * Added di aliases
 * Added markdown engines